update 글 작성시 유저 검증 추가

+++ src/posts/posts.service.ts
@@ -105,7 +105,11 @@ export class PostsService {
   }
 
   // 게시글 삭제
-  async remove(id: number) {
+  async remove(id: number, userId: number) {
+    const editor = await this.matchPostUser(id, userId);
+    if (!editor) {
+      throw new EditorNotMatched();
+    }
     await this.postsRepository.removePost(id);
     return { message: '게시글이 삭제되었습니다' };
   }
